[bugfix] give spliced projections the LM dtype
_build_backbone casts the LM to lm_parameter_dtype, but _build_projections
builds nn.Linear/MLP at the default fp32 and never casts them. index_copy
requires self and source to share a dtype, so build_input raised on the
first forward for any BF16 or FP16 config with a projected slot -- both
dtypes the proto advertises.

The spliced values now take embeds.dtype. Casting the projections
themselves would put their weights in bf16, which is the Adam underflow
the fp32 master-weight default exists to avoid; reading embeds.dtype also
adapts under autocast.

+++ tzrec/models/prompt_generative_model.py
@@ -182,7 +182,7 @@ class BasePromptGenerativeModel(BaseModel):
         return embeds.index_copy(
             0,
             batch.additional_infos[PROMPT_HOLE_POSITIONS],
-            torch.cat(projected_embeddings),
+            torch.cat(projected_embeddings).to(embeds.dtype),
         )
 
     def _tokens_to_local_codes(
